Move all SampleTaskInbox related files into the same folder
