+++ app/packs/src/apps/mydb/collections/CollectionTree.js
@@ -11,8 +11,8 @@ import UIStore from 'src/stores/alt/stores/UIStore';
 import ElementStore from 'src/stores/alt/stores/ElementStore';
 import InboxStore from 'src/stores/alt/stores/InboxStore';
 import UserInfos from 'src/apps/mydb/collections/UserInfos';
-import SampleTaskNavigationElement from 'src/apps/mydb/collections/SampleTaskNavigationElement';
-import SampleTaskInbox from 'src/apps/mydb/collections/SampleTaskInbox';
+import SampleTaskNavigationElement from 'src/apps/mydb/collections/sampleTaskInbox/SampleTaskNavigationElement';
+import SampleTaskInbox from 'src/apps/mydb/collections/sampleTaskInbox/SampleTaskInbox';
 
 import DeviceBox from 'src/apps/mydb/inbox/DeviceBox';
 import UnsortedBox from 'src/apps/mydb/inbox/UnsortedBox';
