+++ app/packs/src/apps/mydb/collections/sampleTaskInbox/SampleTaskInbox.js
