+++ app/packs/src/apps/mydb/collections/sampleTaskInbox/SampleTaskNavigationElement.js
